Adding first snippet
This example of technical writing comes from my undergraduate research project titled "Learning to Move -Simulated Robot Arm Control Through Reinforcement Learning". The specific content of the snippet shows my description of the implementation of a step function which steps the simulation environment to each subsequent time frame in order to train and/or test the reinforcement learning model.